Condense node list styling (#149)
* Condense node selector panel nodes list

* eslint fixes

* Fix eslint settings and node selector styling

* Refactor NodeSelectorPanel a bit

* memo categories & maybe fix eslint temporarily

* remove trailing comma rule from prettier
